Split WalkController into WalkController and WalkStatusController, started working on documentation, WalkStatusController written and tested so far

+++ src/main/java/com/example/petoasisbackend/Service/WalkService.java
@@ -1,6 +1,9 @@
 package com.example.petoasisbackend.Service;
 
 
+import com.example.petoasisbackend.Exception.Walk.WalkDoesntExistException;
+import com.example.petoasisbackend.Exception.WalkStatus.WalkStatusDoesntExistException;
+import com.example.petoasisbackend.Exception.WalkStatus.WalkStatusUpdateCollisionException;
 import com.example.petoasisbackend.Model.Activity.Walk;
 import com.example.petoasisbackend.Model.Descriptor.WalkStatus;
 import com.example.petoasisbackend.Repository.WalkRepository;
@@ -13,24 +16,11 @@ import java.util.List;
 @Service
 public class WalkService {
     @Autowired
-    private WalkStatusRepository walkStatusRepository;
-    @Autowired
     private WalkRepository walkRepository;
-
-    public List<WalkStatus> getWalkStatuses() {
-        return walkStatusRepository.findAll();
-    }
+    private WalkStatusRepository walkStatusRepository;
 
     public List<Walk> getWalks() {
         return walkRepository.findAll();
-    }
-
-
-    public WalkStatus addWalkStatus(WalkStatus status) {
-        if (walkStatusRepository.existsByStatus(status.getStatus())) {
-            throw new IllegalArgumentException(status + " walk status already exists!");
-        }
-        return walkStatusRepository.save(status);
     }
 
 
@@ -38,43 +28,32 @@ public class WalkService {
         return walkRepository.save(walk);
     }
 
-    public Walk removeWalk(Long id) {
+    public Walk removeWalk(Long id) throws WalkDoesntExistException {
+        if (!walkRepository.existsById(id)) {
+            throw new WalkDoesntExistException("Cannot delete walk with id " + id + ". It doesn't exist.");
+        }
         Walk walk = walkRepository.getReferenceById(id);
         walkRepository.delete(walk);
         return walk;
     }
 
-    public Walk updateWalkStatus(Long walkId, String status) {
+    public Walk updateWalkStatus(Long walkId, String status) throws WalkStatusDoesntExistException, WalkDoesntExistException {
         if (!walkStatusRepository.existsByStatus(status)) {
-            throw new IllegalArgumentException(status + " walk status doesnt exist!");
+            throw new WalkStatusDoesntExistException(
+                    "Cannot update walk status of walk with id '" + walkId + "' because status '" +
+                            status + "' doesn't exist"
+            );
+        }
+        if (!walkRepository.existsById(walkId)) {
+            throw new WalkDoesntExistException(
+                    "Cannot update walk status of walk with id '" + walkId + "' because walk with this id doesn't exist"
+            );
         }
         Walk walk = walkRepository.getReferenceById(walkId);
         WalkStatus walkStatus = walkStatusRepository.getWalkStatusByStatus(status);
         walk.setWalkStatus(walkStatus);
         walkRepository.save(walk);
         return walk;
-    }
-
-    public WalkStatus deleteWalkStatus(String statusName) {
-        if (!walkStatusRepository.existsByStatus(statusName)) {
-            throw new IllegalArgumentException(statusName + " walk status doesnt exist!");
-        }
-        WalkStatus status = walkStatusRepository.getWalkStatusByStatus(statusName);
-        walkStatusRepository.delete(status);
-        return status;
-    }
-
-    public WalkStatus updateWalkStatus(String name, String newName) {
-        if (!walkStatusRepository.existsByStatus(name)) {
-            throw new IllegalArgumentException(name + " walk status doesnt exist!");
-        }
-        if (!walkStatusRepository.existsByStatus(newName)) {
-            throw new IllegalArgumentException(name + " walk status already exists!");
-        }
-        WalkStatus status = walkStatusRepository.getWalkStatusByStatus(name);
-        status.setStatus(newName);
-        walkStatusRepository.save(status);
-        return status;
     }
 
 }
